VotingWizard/src/asgn1Election/SimpleElection.java
/**
 * 
 * This file is part of the VotingWizard Project, written as 
 * part of the assessment for CAB302, Semester 1, 2016. 
 * 
 */
package asgn1Election;

import java.util.Iterator;
import asgn1Util.Strings;

/**
 * 
 * Subclass of <code>Election</code>, specialised to simple, first past the post voting
 * 
 * @author hogan
 * 
 */
public class SimpleElection extends Election {

	/**
	 * Simple Constructor for <code>SimpleElection</code>, takes name and also sets the 
	 * election type internally. 
	 * 
	 * @param name <code>String</code> containing the Election name
	 */
	public SimpleElection(String name) {
		super(name);
	}

	/*
	 * (non-Javadoc)
	 * 
	 * @see asgn1Election.Election#findWinner()
	 */
	@Override
	public String findWinner() {
		String result = this.showResultHeader();
		result += "Counting primary votes; " + this.numCandidates + " alternatives available\n";
		this.vc.countPrimaryVotes(this.cds);
		result += this.reportCountResult();
		Candidate winner = this.clearWinner(0);
		
		result += this.reportWinner(winner);
		
		return result;
		}

	/* 
	 * (non-Javadoc)
	 * @see asgn1Election.Election#isFormal(asgn1Election.Vote)
	 */
	@Override
	public boolean isFormal(Vote v) {
		Iterator<Integer> iter = v.iterator();
		int value;
		boolean firstFound = false;
		
		while (iter.hasNext()){
			value = iter.next();
			if(value > this.numCandidates){
				return false;
			}
			if(value == 1){
				if(firstFound){
					return false;
				}
				else{
					firstFound = true;
				}
			}
		}
		if(firstFound){
			return true;
		}
		else{
			return false;
		}
	}

	/*
	 * (non-Javadoc)
	 * 
	 * @see java.lang.Object#toString()
	 */
	@Override
	public String toString() {
		String str = this.name + " - Simple Voting";
		return str;
	}
	
	// Protected and Private/helper methods below///

	/*
	 * (non-Javadoc)
	 * 
	 * @see asgn1Election.Election#clearWinner(int)
	 */
	@Override
	protected Candidate clearWinner(int wVotes) {
		int highestSeen = 0;
		Candidate currentWinner = null;
		
		for(Candidate cand: this.cds.values()){
			if(cand.getVoteCount() > highestSeen){
				highestSeen = cand.getVoteCount();
				currentWinner = cand;
			}
		}
		
		return currentWinner;
	}

	/**
	 * Helper method to create a string reporting the count result
	 * 
	 * @return <code>String</code> containing summary of the count
	 */
	private String reportCountResult() {
		String str = "\nSimple election: " + this.name + "\n\n"
				+ candidateVoteSummary() + "\n";
		String inf = "Informal";
		String voteStr = "" + this.vc.getInformalCount();
		int length = ElectionManager.DisplayFieldWidth - inf.length()
				- voteStr.length();
		str += inf + Strings.createPadding(' ', length) + voteStr + "\n\n";

		String cast = "Votes Cast";
		voteStr = "" + this.numVotes;
		length = ElectionManager.DisplayFieldWidth - cast.length()
				- voteStr.length();
		str += cast + Strings.createPadding(' ', length) + voteStr + "\n\n";
		return str;
	}
}